moved fft implem to a reusable template
