+++ spqlios/ext/neon_accel/macrofx4.h
@@ -1,22 +1,15 @@
 /*
- * 64-bit Floating point NEON macro x4
- *
+ * This file is extracted from the implementation of the FFT on Arm64/Neon
+ * available in https://github.com/cothan/Falcon-Arm (neon/macrof.h).
  * =============================================================================
  * Copyright (c) 2022 by Cryptographic Engineering Research Group (CERG)
  * ECE Department, George Mason University
  * Fairfax, VA, U.S.A.
- * Author: Duc Tri Nguyen
+ * @author: Duc Tri Nguyen dnguye69@gmu.edu, cothannguyen@gmail.com
  * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *     http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
  * =============================================================================
- * @author   Duc Tri Nguyen <dnguye69@gmu.edu>
+ *
+ * This 64-bit Floating point NEON macro x4 has not been modified and is provided as is.
  */
 
 #ifndef MACROFX4_H
