+++ spqlios/reim/reim_fft_neon.c
@@ -1,29 +1,24 @@
 /*
-* High-speed vectorize FFT code for arbitrary `logn`.
-*
-* =============================================================================
-* Copyright (c) 2022 by Cryptographic Engineering Research Group (CERG)
-* ECE Department, George Mason University
-* Fairfax, VA, U.S.A.
-* Author: Duc Tri Nguyen
-* Licensed under the Apache License, Version 2.0 (the "License");
-* you may not use this file except in compliance with the License.
-* You may obtain a copy of the License at
-*     http://www.apache.org/licenses/LICENSE-2.0
-* Unless required by applicable law or agreed to in writing, software
-* distributed under the License is distributed on an "AS IS" BASIS,
-* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-* See the License for the specific language governing permissions and
-* limitations under the License.
-* =============================================================================
-* @author   Duc Tri Nguyen <dnguye69@gmu.edu>
-*/
+ * This file is adapted from the implementation of the FFT on Arm64/Neon
+ * available in https://github.com/cothan/Falcon-Arm (neon/fft.c).
+ * =============================================================================
+ * Copyright (c) 2022 by Cryptographic Engineering Research Group (CERG)
+ * ECE Department, George Mason University
+ * Fairfax, VA, U.S.A.
+ * @author: Duc Tri Nguyen dnguye69@gmu.edu, cothannguyen@gmail.com
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * =============================================================================
+ *
+ * The original source file has been modified by the authors of spqlios-arithmetic
+ * to be interfaced with the dynamic twiddle factors generator of the spqlios-arithmetic
+ * library, as well as the recursive dfs strategy for large complex dimensions m>=2048.
+ */
 
 #include <math.h>
 
 #include "../commons.h"
-#include "macrof.h"
-#include "macrofx4.h"
+#include "../ext/neon_accel/macrof.h"
+#include "../ext/neon_accel/macrofx4.h"
 
 void fill_reim_fft16_omegas_neon(const double entry_pwr, double** omg) {
   const double j_pow = 1. / 8.;
@@ -1599,129 +1594,7 @@ void fill_reim_bitwiddle_fft_omegas_ref(const double rs0, double** omg) {
 #define reim_fft_f reim_fft_neon
 #define fill_reim_fft_omegas_f fill_reim_fft_omegas_neon
 
-void reim_fft_bfs_16_f(uint64_t m, double* re, double* im, double** omg) {
-  uint64_t log2m = log2(m);
-  uint64_t mm = m;
-  if (log2m % 2 != 0) {
-    uint64_t h = mm >> 1;
-    // do the first twiddle iteration normally
-    reim_twiddle_fft_f(h, re, im, *omg);
-    *omg += reim_twiddle_fft_pom_offset;
-    mm = h;
-  }
-  while (mm > 16) {
-    uint64_t h = mm >> 2;
-    for (uint64_t off = 0; off < m; off += mm) {
-      reim_bitwiddle_fft_f(h, re + off, im + off, *omg);
-      *omg += reim_bitwiddle_fft_pom_offset;
-    }
-    mm = h;
-  }
-  if (mm != 16) abort();  // bug!
-  for (uint64_t off = 0; off < m; off += 16) {
-    reim_fft16_f(re + off, im + off, *omg);
-    *omg += reim_fft16_pom_offset;
-  }
-}
-
-void fill_reim_fft_bfs_16_omegas_f(uint64_t m, double entry_pwr, double** omg) {
-  uint64_t log2m = log2(m);
-  uint64_t mm = m;
-  double ss = entry_pwr;
-  if (log2m % 2 != 0) {
-    uint64_t h = mm >> 1;
-    double s = ss / 2.;
-    // do the first twiddle iteration normally
-    fill_reim_twiddle_fft_omegas_f(s, omg);
-    mm = h;
-    ss = s;
-  }
-  while (mm > 16) {
-    uint64_t h = mm >> 2;
-    double s = ss / 4.;
-    for (uint64_t off = 0; off < m; off += mm) {
-      double rs0 = s + fracrevbits(off / mm) / 4.;
-      fill_reim_bitwiddle_fft_omegas_f(rs0, omg);
-    }
-    mm = h;
-    ss = s;
-  }
-  if (mm != 16) abort();  // bug!
-  for (uint64_t off = 0; off < m; off += 16) {
-    double s = ss + fracrevbits(off / 16);
-    fill_reim_fft16_omegas_f(s, omg);
-  }
-}
-
-void reim_fft_rec_16_f(uint64_t m, double* re, double* im, double** omg) {
-  if (m <= 2048) return reim_fft_bfs_16_f(m, re, im, omg);
-  const uint32_t h = m / 2;
-  reim_twiddle_fft_f(h, re, im, *omg);
-  *omg += reim_twiddle_fft_pom_offset;
-  reim_fft_rec_16_f(h, re, im, omg);
-  reim_fft_rec_16_f(h, re + h, im + h, omg);
-}
-
-void fill_reim_fft_rec_16_omegas_f(uint64_t m, double entry_pwr, double** omg) {
-  if (m <= 2048) return fill_reim_fft_bfs_16_omegas_f(m, entry_pwr, omg);
-  const uint64_t h = m / 2;
-  const double s = entry_pwr / 2;
-  fill_reim_twiddle_fft_omegas_f(s, omg);
-  fill_reim_fft_rec_16_omegas_f(h, s, omg);
-  fill_reim_fft_rec_16_omegas_f(h, s + 0.5, omg);
-}
-
-void reim_fft_f(const REIM_FFT_PRECOMP* precomp, double* dat) {
-  const int32_t m = precomp->m;
-  double* omg = precomp->powomegas;
-  double* re = dat;
-  double* im = dat + m;
-  if (m <= 16) {
-    switch (m) {
-      case 1:
-        return;
-      case 2:
-        return reim_fft2_f(re, im, omg);
-      case 4:
-        return reim_fft4_f(re, im, omg);
-      case 8:
-        return reim_fft8_f(re, im, omg);
-      case 16:
-        return reim_fft16_f(re, im, omg);
-      default:
-        abort();  // m is not a power of 2
-    }
-  }
-  if (m <= 2048) return reim_fft_bfs_16_f(m, re, im, &omg);
-  return reim_fft_rec_16_f(m, re, im, &omg);
-}
-
-void fill_reim_fft_omegas_f(uint64_t m, double entry_pwr, double** omg) {
-  if (m <= 16) {
-    switch (m) {
-      case 1:
-        break;
-      case 2:
-        fill_reim_fft2_omegas_f(entry_pwr, omg);
-        break;
-      case 4:
-        fill_reim_fft4_omegas_f(entry_pwr, omg);
-        break;
-      case 8:
-        fill_reim_fft8_omegas_f(entry_pwr, omg);
-        break;
-      case 16:
-        fill_reim_fft16_omegas_f(entry_pwr, omg);
-        break;
-      default:
-        abort();  // m is not a power of 2
-    }
-  } else if (m <= 2048) {
-    fill_reim_fft_bfs_16_omegas_f(m, entry_pwr, omg);
-  } else {
-    fill_reim_fft_rec_16_omegas_f(m, entry_pwr, omg);
-  }
-}
+#include "reim_fft_core_template.h"
 
 EXPORT REIM_FFT_PRECOMP* new_reim_fft_precomp_neon(uint32_t m, uint32_t num_buffers) {
   const uint64_t OMG_SPACE = ceilto64b(2 * m * sizeof(double));
